Update waveform_generator.py
When `mode_list` is specified in a WaveformGenerator instance, lal_params is set to a lal_dict object, which can not be pickled.

Changed the WaveformGenerator class to include mode_list as a an attribute and changed `lal_params = self.setup_mode_array(mode_list)` to be called inside `_convert_parameters()` instead of being passed to it.  This avoids a lal_dict object being stored in the WaveformGenerator, thus allowing for parallel waveform generation.